small changes, map nearly complete. Units loaded into clientside, still need to be added to the unit layer. Added admin access level for editing inspectors. Loaded inspector's to edit, will show a button if any inspectors come back from the edit endpoint. Still need to add inspector edit menu Still need to add grouping circle around locations that have multiple inspectors assigned.

+++ ClayInspectionView/Models/Unit.cs
@@ -9,8 +9,8 @@ namespace ClayInspectionView.Models
   {
     public string Name { get; set; }
     public DateTime Date_Last_Communicated { get; set; }
-    public double Longitude { get; set; }
-    public double Latitude { get; set; }
+    public decimal Longitude { get; set; }
+    public decimal Latitude { get; set; }
     public string Unit_Icon_URL
     {
       get
